% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{collect_data}
\alias{collect_data}
\title{Read data using SQL scripts}
\usage{
collect_data(scripts, con = connect_to_database())
}
\arguments{
\item{scripts}{A list of character vectors containing SQL}

\item{con}{A connection to an SQL database}
}
\value{
A named list of data.frames
}
\description{
Read data using SQL scripts
}
