% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Connect to a database}
\usage{
connect_to_database(conf = config::get("database"))
}
\arguments{
\item{conf}{List of config opts}
}
\value{
A connection
}
\description{
This function attempts to generate a connection using \code{DBI::dbConnect()}. By
default, options are taken from a \code{config.yml} file which should contain
fields for
\itemize{
\item driver
\item server
\item database
\item uid (username)
\item pwd (password)
\item trusted (usually 'yes')
}
}
