% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Create a project for a new report}
\usage{
create_report(
  path,
  excel_files = getOption("qr.default_template"),
  r_files = NULL,
  sql_files = NULL,
  rstudio = rstudioapi::isAvailable(),
  open = interactive()
)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{excel_files, r_files, sql_files}{Optional directories containing
pre-existing files to use in the report}

\item{rstudio}{If \code{TRUE}, calls \code{use_rstudio()} to make the new package or
project into an RStudio Project. If FALSE and a non-package project, a
sentinel .here file is placed so that the directory can be recognized as a
project by the here or rprojroot packages.}

\item{open}{If \code{TRUE}, activates the new project:
\itemize{
\item If RStudio desktop, the package is opened in a new session.
\item If on RStudio server, the current RStudio project is activated.
\item Otherwise, the working directory and active project is changed.
}}
}
\value{
Path to the newly created project or package, invisibly.
}
\description{
This function works very similarly to \code{usethis::create_project()}, but sets
up the project to generate an Excel report using {quickReport}
}
