% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{default_template}
\alias{default_template}
\title{Default Excel report template}
\format{
An object of class \code{Workbook} of length 1.
}
\usage{
default_template
}
\description{
This is a workbook object containing a (very) basic coversheet template
}
\keyword{datasets}
