% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_db.R
\name{example_db}
\alias{example_db}
\title{Create an example SQLite database}
\usage{
example_db()
}
\value{
A connection
}
\description{
This function creates an SQLite database in memory containing 3 base
datasets: \code{mtcars}, \code{beaver1} and \code{C02}.
}
