% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-utils.R
\name{get_user}
\alias{get_user}
\title{Get the current user}
\usage{
get_user(default = getOption("qr.user"))
}
\arguments{
\item{default}{A default which can be set using \code{options()}}
}
\value{
A character vector of length 1
}
\description{
This function gets the name of the current user, which is used to set the
'Creator' and 'Last Modified By' fields when running \code{report_from_sql()} or
\code{report_from_data()}
}
\details{
If a \code{default} is not set this function gets the username from
\itemize{
\item \code{Sys.getenv("USERNAME")} if on windows
\item \code{Sys.getenv("USER")} otherwise
\item If neither are found the \code{"Unknown"} is returned.
}
}
