% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-utils.R
\name{info_table}
\alias{info_table}
\alias{default_info_table}
\title{Create a pretty table of information}
\usage{
info_table(..., .pretty_names = TRUE)

default_info_table(
  data_included = "",
  data_exported_by = get_user(),
  exported_on = strftime(Sys.time(), "\%Y-\%m-\%d"),
  .pretty_names = TRUE,
  ...
)
}
\arguments{
\item{...}{Named vectors. The names give the single value that will appear in
the left-hand column, while the values will appear on separate rows in the
right-hand column.}

\item{.pretty_names}{If \code{TRUE} then some standardisation is performed on the
values in the left-hand column. This includes the capitalisation of the
first letter and the replacement of underscores with spaces.}

\item{data_included}{The names of the data sheets included in the report}

\item{data_exported_by}{The name of the reporter}

\item{exported_on}{The date of the report}
}
\value{
A \code{data.frame}
}
\description{
These functions help you create a readable table of information, intended to
be used to populate a coversheet in an Excel report.
}
\examples{
info_table(
  first_category = letters[1:3],
  second_category = "d"
)

default_info_table(
  data_included = c("Blackberry", "Apple", "Orange"),
  author = "Jacob"
)
}
