% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-files.R
\name{nice_filename}
\alias{nice_filename}
\title{Generate a filename}
\usage{
nice_filename(
  name = NULL,
  user = Sys.info()[["user"]],
  time = strftime(Sys.time(), "\%Y-\%m-\%d"),
  ext = ".xlsx",
  dir = NULL,
  replace = FALSE
)
}
\arguments{
\item{name}{A brief description of the file}

\item{user}{The name of the creater of the file}

\item{time}{The time the file was created (i.e. \code{Sys.time()})}

\item{ext}{The file extension}

\item{dir}{The directory to save the file to}

\item{replace}{If \code{FALSE} then filenames that already exist will have a number
appended to ensure that replacement doesn't occur.}
}
\value{
A character vector of length 1
}
\description{
This is a helper function to quickly generate not-terrible filenames. This
will be of the form 'day created-author-description.xlsx'
Filenames are put in lower-case and spaces are replaced with dashes.
}
\examples{
nice_filename("demo_file", user = "example")
}
