% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_from_.R
\name{populate_coversheet}
\alias{populate_coversheet}
\title{Populate the report coversheet}
\usage{
populate_coversheet(
  wb = initialise_report(),
  data = list(),
  sheet = "Cover Sheet",
  ...
)
}
\arguments{
\item{wb}{A workbook generated by \code{initialise_report()}}

\item{data}{A named list of data to write. Names give cell entries}

\item{sheet}{The name of the sheet to write to}

\item{...}{Extra options passed to \code{openxlsx::writeData()}}
}
\description{
Populate the report coversheet
}
