% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-scripts.R
\name{read_sql_scripts}
\alias{read_sql_scripts}
\title{Read the SQL scripts in the SQL directory}
\usage{
read_sql_scripts(dir = "SQL", engine = readr::read_file)
}
\arguments{
\item{dir}{The SQL directory}

\item{engine}{Function used to read the scripts}
}
\value{
A list of character vectors
}
\description{
Read the SQL scripts in the SQL directory
}
