% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_from_.R
\name{save_excel_export}
\alias{save_excel_export}
\title{Save an Excel export}
\usage{
save_excel_export(wb, filepath, replace = FALSE)
}
\arguments{
\item{wb}{The workbook to save}

\item{filepath}{The filepath to save to}

\item{replace}{If \code{TRUE} then the function will fail if such a file already
exists.}
}
\description{
Save an Excel export
}
