% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_tables}
\alias{transform_tables}
\title{Apply functions to a list of data.frames}
\usage{
transform_tables(tables, trans)
}
\arguments{
\item{tables}{A named list of data.frames}

\item{trans}{Can be either a single function that will be applied to each
element of \code{tables} or a named list of functions. If the latter:
\itemize{
\item Names should correspond to the names of \code{tables} to indicate the tables
the functions should be applied to
\item If you supply an unnamed function, this will be applied to each table
that doesn't have a corresponding named function. You can think of this as
a 'default' transformation.
}}
}
\value{
A list of data.frames
}
\description{
This function provides an interface for applying functions to a named
list of data.frames.
}
