% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming-files.R
\name{validate_filename}
\alias{validate_filename}
\title{Get a valid filepath}
\usage{
validate_filename(
  filename = nice_filename(),
  dir = NULL,
  replace = FALSE,
  filepath = paste(c(dir, filename), collapse = "/")
)
}
\arguments{
\item{filename}{The name of the file to check}

\item{dir}{The directory}

\item{replace}{If \code{TRUE} then the input is returned regardless of whether
the file exists}

\item{filepath}{Optionally supply the full filepath instead of \code{filename} and
\code{dir}}
}
\value{
A character string
}
\description{
Get a valid filepath
}
