% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data_sheets.R
\name{write_data_sheets}
\alias{write_data_sheets}
\title{Add datasheets to a Workbook}
\usage{
write_data_sheets(
  wb,
  datasets,
  default_data_sheet = 2,
  data_marker = getOption("qr.marker", "!data"),
  remove_template = TRUE
)
}
\arguments{
\item{wb}{A workbook object}

\item{datasets}{A list of datasets}

\item{default_data_sheet}{The worksheet to use as the default datasheet
template. Can be either the sheet name or number.}

\item{data_marker}{This value should be used in the datasheet template to
indicate where the tables should be placed. This value will be the
upper-left corner of the table.}

\item{remove_template}{Logical, whether the template sheet should be removed
from the final report. Does not have an effect if \code{default_data_sheet}
is \code{NULL}.}
}
\value{
An \code{openxlsx} workbook object
}
\description{
Add datasheets to a Workbook
}
