% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-utils.R
\name{read_worksheets}
\alias{read_worksheets}
\title{Helper for reading an entire .xlsx file}
\usage{
read_worksheets(file, ...)
}
\arguments{
\item{file}{A filepath}

\item{...}{Passed to \code{openxlsx::readWorkbook()}}
}
\value{
A list of data.frames
}
\description{
Helper for reading an entire .xlsx file
}
